Changed mib slave code to automatically generate parameter lists for commands that it knows based on a packed structure describing the parameters that the command takes.  This means the command handlers do not have to worry about creating their own parameter lists anymore.
Also removed references to gsm, registration and uart since those are going to be changing
and moved them to an old folder for safekeeping.

+++ mainboard-firmware/src/modules/memory.c
@@ -207,8 +207,10 @@ void mem_clear_all() {
   DISABLE_MEMORY();
 }
 
-void mem_clear_subsection( unsigned int addr ) {
+void mem_clear_subsection( unsigned long addr ) 
+{
   addr &= MEMORY_ADDRESS_MASK;
+  
   mem_wait_while_writing();
   mem_enable_write();
